// webapp/src/types/GetEntityItemsResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EntityItemWithHighlights } from "./EntityItemWithHighlights";

export type GetEntityItemsResponse = { items: Array<EntityItemWithHighlights>, total_count: number, page: number, limit: number, total_pages: number, };